+++ src/modules/block.js
@@ -53,9 +53,23 @@ module.exports = ({ bot, knex, config, commands }) => {
 
     if (expiresAt) {
       const humanized = humanizeDuration(args.blockTime, { largest: 2, round: true });
-      channel.createMessage(`Blocked <@${userIdToBlock}> (id \`${userIdToBlock}\`) from modmail for ${humanized}`);
+      msg.channel.createMessage(`Blocked <@${userIdToBlock}> (id \`${userIdToBlock}\`) from modmail for ${humanized}`);
+
+      const timedBlockMessage = config.timedBlockMessage || config.blockMessage;
+      if (timedBlockMessage) {
+        const dmChannel = await user.getDMChannel();
+        const formatted = timedBlockMessage
+          .replace(/\{duration}/g, humanized)
+          .replace(/\{timestamp}/g, moment.utc(expiresAt).format("X"));
+        dmChannel.createMessage(formatted).catch(utils.noop);
+      }
     } else {
-      channel.createMessage(`Blocked <@${userIdToBlock}> (id \`${userIdToBlock}\`) from modmail indefinitely`);
+      msg.channel.createMessage(`Blocked <@${userIdToBlock}> (id \`${userIdToBlock}\`) from modmail indefinitely`);
+
+      if (config.blockMessage != null) {
+        const dmChannel = await user.getDMChannel();
+        dmChannel.createMessage(config.blockMessage).catch(utils.noop);
+      }
     }
   };
 
@@ -81,9 +95,23 @@ module.exports = ({ bot, knex, config, commands }) => {
       const humanized = humanizeDuration(args.unblockDelay, { largest: 2, round: true });
       await blocked.updateExpiryTime(userIdToUnblock, unblockAt);
       msg.channel.createMessage(`Scheduled <@${userIdToUnblock}> (id \`${userIdToUnblock}\`) to be unblocked in ${humanized}`);
+
+      const timedUnblockMessage = config.timedUnblockMessage || config.unblockMessage;
+      if (timedUnblockMessage) {
+        const dmChannel = await user.getDMChannel();
+        const formatted = timedUnblockMessage
+          .replace(/\{delay}/g, humanized)
+          .replace(/\{timestamp}/g, moment.utc(unblockAt).format("X"))
+        dmChannel.createMessage(formatted).catch(utils.noop);
+      }
     } else {
       await blocked.unblock(userIdToUnblock);
       msg.channel.createMessage(`Unblocked <@${userIdToUnblock}> (id ${userIdToUnblock}) from modmail`);
+
+      if (config.unblockMessage) {
+        const dmChannel = await user.getDMChannel();
+        dmChannel.createMessage(config.unblockMessage).catch(utils.noop);
+      }
     }
   };